Refactor accessibility utilities and update imports for consistency
- Moved accessibility-related functions to a new `spectrum/index.js` module, enhancing modularity and organization.
- Replaced direct imports from the old accessibility utility file with the new module across various components, ensuring consistent usage.
- Deleted the outdated accessibility and overlay utility files to streamline the codebase and reduce redundancy.
- Introduced a new `loadCss.js` utility for improved CSS loading management, enhancing performance and maintainability.
